+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## [0.4] - Unreleased
+## [0.4] - 2024-04-05
 
 - The row count shown in the list-of-tables buffer is now precise even when the tables have not been
   VACUUMed. This precision is at the cost of speed on large tables (calculated with `COUNT(*)`).